Move and rename the shared adapter spec to stop warnings
