+++ spec/models/adapters/active_record_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 require_relative "../../../app/models/adapters/active_record"
-require_relative "../../support/shared/adapter_spec"
+require_relative "../../shared/adapter"
 
 RSpec.describe Adapters::ActiveRecord do
   it_behaves_like "an adapter"
