+++ spec/shared/adapter.rb
